requirements.txt
pandas>=1.5
numpy>=1.23
requests>=2.28
python-dotenv>=1.0
pytz
langchain>=0.2.14
langchain-core>=0.2.33
langchain-community>=0.2.12
langchain-openai>=0.1.23
langgraph>=0.2.32
faiss-cpu
tqdm
openai>=1.40.0
sentence-transformers
scikit-learn
fastapi
uvicorn